requirements.txt
Django==1.4
argparse==1.2.1
dj-database-url==0.2.1
django-recaptcha==0.0.4
django-registration==0.8
django-social-auth==0.7.5
django-tagging==0.3.1
django-tagging-autocomplete==0.3.1
django-taggit==0.9.3
facepy==0.8.2
httplib2==0.7.2
oauth2==1.5.211
psycopg2==2.4.5
python-openid==2.2.5
requests==0.14.1
simplejson==2.3.2
virtualenv==1.7.1.2
wsgiref==0.1.2
